+++ R/data-corpus_rae.R
@@ -1,0 +1,21 @@
+#' @title Corpus de Referencia del Español Actual (CREA) escrito
+#'
+#' @description
+#' Un corpus de referencia ha de ser lo suficientemente extenso para representar
+#' todas las variedades relevantes de una lengua. En la versión 3.2, cerrada en junio de 2008
+#' y consultable en la actualidad, la parte escrita del CREA cuenta con casi 140 000 documentos
+#' y algo más de 154 millones de formas procedentes de textos de todos los países
+#' hispánicos y producidos entre 1975 y 2004.
+#'
+#' @format Un data.frame con 3 variables
+#'
+#' @source \url{https://www.rae.es/recursos/banco-de-datos/crea-escrito}
+#'
+#' \describe{
+#'   \item{Orden}{formas ortográficas registradas en el CREA}
+#'   \item{Frec.absoluta}{frecuencia absoluta de la palabra}
+#'   \item{Frec.normalizada}{frecuencia normalizada de la palabra}
+#' }
+#'
+
+"regiones_y_comunas"