+++ tests/testthat/test-rec_ortografia.R
@@ -1,0 +1,4 @@
+test_that("multiplication works", {
+  expect_equal(rec_ortografia(c('pais', 'arbol')),
+                              c('país', 'árbol'))
+})