+++ NEWS.md
@@ -7,6 +7,7 @@
 * Función `is_email` para validar sintaxis de correos electrónicos.
 * Función `svy_tabla_vars_segmentos` para obtener estadísticos de multiples variables y segmentos.
 * Función `fct_case_when` a partir de @pewmethods.
+* funcion `rec_ortografia` para corregir ortografía de palabras.
 
 ## Features